fs: Give initialisaiton code it's own file
